+++ R/gg-plots.R
@@ -11,17 +11,58 @@ if (getRversion() >= "2.15.1") {
 
 
 # retrieve the evaulated data column given the aes (which could possibly do operations)
+#' Evaluate data column
+#' @param data data set to evaluate the data with
+#' @param aes_col Single value from an \code{ggplot2::\link[ggplot2]{aes}(...)} object
+#' @return Aes mapping with the x and y values switched
+#' @export
+#' @examples
+#' mapping <- ggplot2::aes(Petal.Length)
+#' eval_data_col(iris, mapping$x)
 eval_data_col <- function(data, aes_col) {
-  eval(aes_col, data)
+  rlang::eval_tidy(aes_col, data)
+}
+#' Aes name
+#' @param aes_col Single value from \code{ggplot2::\link[ggplot2]{aes}(...)}
+#' @return character string
+#' @export
+#' @examples
+#' mapping <- ggplot2::aes(Petal.Length)
+#' mapping_string(mapping$x)
+mapping_string <- function(aes_col) {
+  gsub("^~", "", deparse(aes_col, 500L))
 }
 
 # is categories on the left?
-is_character_column <- is_horizontal <- function(data, mapping, val = "y") {
+#' Check if plot is horizontal
+#'
+#' @param data data used in ggplot2 plot
+#' @param mapping ggplot2 \code{aes()} mapping
+#' @param val key to retrieve from \code{mapping}
+#' @return Boolean determining if the data is a character-like data
+#' @export
+#' @rdname is_horizontal
+#' @examples
+#' is_horizontal(iris, ggplot2::aes(Sepal.Length, Species)) # TRUE
+#' is_horizontal(iris, ggplot2::aes(Sepal.Length, Species), "x") # FALSE
+#' is_horizontal(iris, ggplot2::aes(Sepal.Length, Sepal.Width)) # FALSE
+is_horizontal <- function(data, mapping, val = "y") {
   yData <- eval_data_col(data, mapping[[val]])
 
   is.factor(yData) || is.character(yData)
 }
+#' @export
+#' @rdname is_horizontal
+is_character_column <- is_horizontal
 
+#' Swap x and y mapping
+#' @param mapping output of \code{ggplot2::\link[ggplot2]{aes}(...)}
+#' @return Aes mapping with the x and y values switched
+#' @export
+#' @examples
+#' mapping <- ggplot2::aes(Petal.Length, Sepal.Width)
+#' mapping
+#' mapping_swap_x_y(mapping)
 mapping_swap_x_y <- function(mapping) {
   tmp <- mapping$x
   mapping$x <- mapping$y
@@ -247,39 +288,25 @@ ggally_cor <- function(
 
   # mapping$x <- mapping$y <- NULL
 
-  xCol <- deparse(mapping$x)
-  yCol <- deparse(mapping$y)
+  xData <- eval_data_col(data, mapping$x)
+  yData <- eval_data_col(data, mapping$y)
 
-  if (is_date(data[[xCol]]) || is_date(data[[yCol]])) {
-
-    # make sure it's a data.frame, as data.tables don't work well
-    if (! identical(class(data), "data.frame")) {
-      data <- fix_data(data)
-    }
-
-    for (col in c(xCol, yCol)) {
-      if (is_date(data[[col]])) {
-        data[[col]] <- as.numeric(data[[col]])
-      }
-    }
+  if (is_date(xData)) {
+    xData <- as.numeric(xData)
   }
-
-  if (is.numeric(eval_data_col(data, mapping$colour))) {
+  if (is_date(yData)) {
+    yData <- as.numeric(yData)
+  }
+  colorData <- eval_data_col(data, mapping$colour)
+  if (is.numeric(colorData)) {
     stop("ggally_cor: mapping color column must be categorical, not numeric")
   }
 
-  colorCol <- deparse(mapping$colour)
-  singleColorCol <- ifelse(is.null(colorCol), NULL, paste(colorCol, collapse = ""))
-
   if (use %in% c("complete.obs", "pairwise.complete.obs", "na.or.complete")) {
-    if (length(colorCol) > 0) {
-      if (singleColorCol %in% colnames(data)) {
-        rows <- complete.cases(data[c(xCol, yCol, colorCol)])
-      } else {
-        rows <- complete.cases(data[c(xCol, yCol)])
-      }
+    if (!is.null(colorData) && (length(colorData) == length(xData))) {
+      rows <- complete.cases(xData, yData, colorData)
     } else {
-      rows <- complete.cases(data[c(xCol, yCol)])
+      rows <- complete.cases(xData, yData)
     }
 
     if (any(!rows)) {
@@ -290,54 +317,64 @@ ggally_cor <- function(
         warning("Removing 1 row that contained a missing value")
       }
     }
-    data <- data[rows, ]
+
+    if (!is.null(colorData) && (length(colorData) == length(xData))) {
+      colorData <- colorData[rows]
+    }
+    xData <- xData[rows]
+    yData <- yData[rows]
   }
 
-  xVal <- data[[xCol]]
-  yVal <- data[[yCol]]
+  xVal <- xData
+  yVal <- yData
 
-  if (length(names(mapping)) > 0){
-    for (i in length(names(mapping)):1){
-      # find the last value of the aes, such as cyl of as.factor(cyl)
-      tmp_map_val <- deparse(mapping[names(mapping)[i]][[1]])
-      if (tmp_map_val[length(tmp_map_val)] %in% colnames(data))
-        mapping[[names(mapping)[i]]] <- NULL
+  # if the mapping has to deal with the data, remove it
+  mappingCopy <- mapping
+  if (packageVersion("ggplot2") > "2.2.1") {
+    for (mappingName in names(mapping)) {
+      itemData <- eval_data_col(data, mapping[[mappingName]])
+      if (!inherits(itemData, "AsIs")) {
+        mapping[[mappingName]] <- NULL
+      }
+    }
+  } else {
+    if (length(names(mapping)) > 0){
+      for (i in length(names(mapping)):1){
+        # find the last value of the aes, such as cyl of as.factor(cyl)
+        tmp_map_val <- deparse(mapping[names(mapping)[i]][[1]])
+        if (tmp_map_val[length(tmp_map_val)] %in% colnames(data))
+          mapping[[names(mapping)[i]]] <- NULL
 
-      if (length(names(mapping)) < 1){
-        mapping <- NULL
-        break;
+        if (length(names(mapping)) < 1){
+          mapping <- NULL
+          break;
+        }
       }
     }
   }
 
-
-  # splits <- str_c(
-  #   as.character(mapping$group), as.character(mapping$colour),
-  #   sep = ", ", collapse = ", "
-  # )
-  # splits <- str_c(colorCol, sep = ", ", collapse = ", ")
-  if (length(colorCol) < 1) {
-    colorCol <- "ggally_NO_EXIST"
-  }
-
   if (
-    (singleColorCol != "ggally_NO_EXIST") &&
-    (singleColorCol %in% colnames(data))
+    !is.null(colorData) &&
+    !inherits(colorData, "AsIs")
   ) {
 
-    cord <- ddply(data, c(colorCol), function(x) {
-      cor_fn(x[[xCol]], x[[yCol]])
-    })
-    colnames(cord)[2] <- "ggally_cor"
+    cord <- ddply(
+      data.frame(x = xData, y = yData, color = colorData),
+      "color",
+      function(dt) {
+        cor_fn(dt$x, dt$y)
+      }
+    )
+    colnames(cord)[2] <- "correlation"
 
-    cord$ggally_cor <- signif(as.numeric(cord$ggally_cor), 3)
+    cord$correlation <- signif(as.numeric(cord$correlation), 3)
 
     # put in correct order
-    lev <- levels(data[[colorCol]])
+    lev <- levels(as.factor(colorData))
     ord <- rep(-1, nrow(cord))
     for (i in 1:nrow(cord)) {
       for (j in seq_along(lev)){
-        if (identical(as.character(cord[i, colorCol]), as.character(lev[j]))) {
+        if (identical(as.character(cord$color[i]), as.character(lev[j]))) {
           ord[i] <- j
         }
       }
@@ -346,8 +383,7 @@ ggally_cor <- function(
     # print(order(ord[ord >= 0]))
     # print(lev)
     cord <- cord[order(ord[ord >= 0]), ]
-
-    cord$label <- str_c(cord[[colorCol]], ": ", cord$ggally_cor)
+    cord$label <- str_c(cord$color, ": ", cord$correlation)
 
     # calculate variable ranges so the gridlines line up
     xmin <- min(xVal, na.rm = TRUE)
@@ -533,7 +569,7 @@ ggally_dot_and_box <- function(data, mapping, ..., boxPlot = TRUE){
     mapping <- mapping_swap_x_y(mapping)
   }
 
-  xVal <- deparse(mapping$x)
+  xVal <- mapping_string(mapping$x)
   mapping$x <- 1
 
   p <- ggplot(data = data)
@@ -621,8 +657,8 @@ ggally_facethist <- function(data, mapping, ...){
     mapping <- mapping_swap_x_y(mapping)
   }
 
-  xVal <- deparse(mapping$x)
-  yVal <- deparse(mapping$y)
+  xVal <- mapping_string(mapping$x)
+  yVal <- mapping_string(mapping$y)
   mapping$y <- NULL
 
   p <- ggplot(data = data, mapping)
@@ -710,8 +746,9 @@ ggally_facetdensitystrip <- function(data, mapping, ..., den_strip = FALSE){
     mapping <- mapping_swap_x_y(mapping)
   }
 
-  xVal <- deparse(mapping$x)
-  yVal <- deparse(mapping$y)
+  xVal <- mapping_string(mapping$x)
+  yVal <- mapping_string(mapping$y)
+  mappingY <- mapping$y
   mapping$y <- NULL # will be faceted
 
   p <- ggplot(data = data, mapping) + labs(x = xVal, y = yVal)
@@ -903,11 +940,9 @@ ggally_text <- function(
 
   # dont mess with color if it's already there
   if (!is.null(mapping$colour)) {
-    colour <- mapping$colour
-    # remove colour from the aesthetics, so legend isn't printed
-    mapping$colour <- NULL
     p <- p +
-       geom_text( label = label, mapping = mapping, colour = colour, ...)
+       geom_text( label = label, mapping = mapping, ...) +
+       guides(colour = FALSE)
   } else if ("colour" %in% names(aes(...))) {
     p <- p +
        geom_text( label = label, mapping = mapping, ...)
@@ -1026,7 +1061,7 @@ ggally_diagAxis <- function(
   numer <- ! is_horizontal(data, mapping, "x")
 
   if (! is.character(label)) {
-    label <- deparse(mapping$x)
+    label <- mapping_string(mapping$x)
   }
 
   xData <- eval_data_col(data, mapping$x)
@@ -1134,7 +1169,7 @@ ggally_facetbar <- function(data, mapping, ...){
 
   # numer <- is.null(attributes(data[,as.character(mapping$x)])$class)
   # xVal <- mapping$x
-  yVal <- deparse(mapping$y)
+  yVal <- mapping_string(mapping$y)
   mapping$y <- NULL
   p <- ggplot(data, mapping) +
     geom_bar(...) +
@@ -1174,8 +1209,8 @@ ggally_ratio <- function(
 ) {
 
   # capture the original names
-  xName <- deparse(mapping$x)
-  yName <- deparse(mapping$y)
+  xName <- mapping_string(mapping$x)
+  yName <- mapping_string(mapping$y)
 
   countData <- plyr::count(data, vars = c(xName, yName))