+++ R/utils.R
@@ -12,16 +12,15 @@ print_if_interactive <- function(p) {
 }
 
 
-#' Require packages
+#' Loads package namespaces
 #'
-#' Requires packages or yells at user... loudly
+#' Loads package namespaces or yells at user... loudly
 #'
 #' @param pkgs vector of character values
-#' @author Barret Schloerke \email{schloerke@@gmail.com}
 #' @keywords internal
-require_pkgs <- function(pkgs) {
+require_namespaces <- function(pkgs) {
   for (pkg in pkgs) {
-    if (! require(pkg, character.only = TRUE)) {
+    if (! requireNamespace(pkg, quietly = TRUE)) {
       stop(str_c("please install the package '", pkg, "'.  install.packages('", pkg, "') "))
     }
   }